+++ lib/core_extensions/all.rb
@@ -4,7 +4,10 @@
 
 require "core_extensions/array"
 require "core_extensions/hash"
+require "core_extensions/true_class"
+require "core_extensions/false_class"
 require "core_extensions/nil_class"
 
-require "core_extensions/object"
+require "core_extensions/class"
 require "core_extensions/module"
+require "core_extensions/object"