+++ lib/core_extensions/nil_class.rb
@@ -3,12 +3,21 @@
 # -*- warn_indent: true -*-
 
 class NilClass
-  # Returns +true+ when invoked on `nil`.
+  # Returns +true+ if invoked on `NilClass` instance.
   #
   # ==== Examples
   #
   #   nil.blank? #=> true
   def blank?
     true
+  end
+
+  # Returns +false+ if invoked on +NilClass+ instance.
+  #
+  # ==== Examples
+  #
+  #   nil.to_b #=> false
+  def to_b
+    false
   end
 end